Added test for xml
